Set inherited flag for properties where applicable
There's an implicit argument to setPropRec that specifies whether
the function is recursing. If it's recursing, that means that
we're looking at a child and therefore the property is inherited.

For verbs, this sort of witchcraft does not need to be used because
we can simply add a different verb directly to the children