Fix egregious error in persist.nim (see inside)
To atomicize the process of persisting the world, I create a temporary
directory to write the world to, which is then copied over to the real
directory. However, as the old code shows, I, for some reason that
completely escapes me, decided that it was a good idea to delete the new
directory right before copying it again (and not forget to delete it one
more time, for good measure). This has cost me a few tenths of an hour
in re-writing verb code that I thought had persisted.