Better task error reporting and more public logging
Made the logger a global variable so anyone can "import logging" and
info/warn away.

Task error backtraces should be a bit more accurate now.

If a task raises an real exception, it should be turned into an
E_INTERNAL and the server should issue a warning. In the end I hope to
abolish all possibilities of exceptions.