Setting owner property should set 'c' prop owners
Cryptic commit message. Here's an explanation:

Imagine you have an object "dog" with property "breed", with the
'c' bit set in permissions. This object is owned by me, and so
is the property. One day, however, I die, so this object (and all
my other possessions) are re-owned to my son. This is done with

  (setprop obj "owner" son)

Now any property on those objects with the 'c' bit set need to
have their owner set to the son. That is what this change
implements. The 'setprop' builtin has a hard-coded hook on the
'owner' property to set the owner of any 'c' properties of
the object being re-owned.

In addition, any attempt to set a builtin property to a value
of type that it wasn't intended to be will result in an error
instead of a silent failure.